test: extract reusable test cases
