+++ tests/types.ts
@@ -1,0 +1,6 @@
+export type TestResult = { name: string; ok: boolean; details: string };
+
+export type TestCase = {
+  name: string;
+  run: () => { ok: boolean; details: string };
+};
